Added reset and pause functionality + refactored
A lot of resources were moved to the Tetris game class so that all the
other classes can access them. The main menu assets were renamed ui to
signify that other classes are using them for ui purposes and not just
the main menu class. Game can now reset and also pause.